pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "flask-auth-service"
version = "0.1.0"
description = "A simple Flask authentication service with Bearer token validation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "David ALEXANDRE", email = "david.alexandre@w6d.io"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Flask",
]
dependencies = [
    "Flask>=2.0.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/w6d-io/flask-auth-service"
"Bug Reports" = "https://github.com/w6d-io/flask-auth-service/issues"
"Source" = "https://github.com/w6d-io/flask-auth-service"

[project.scripts]
flask-auth-service = "flask_auth_service.cli:main"